A new redone version
I was done with the regular assignment within a day, so I decided to redo the given files (along with the new additions). In this version, a lot of the functionality is moved into new classes, in order to make the coding more readable, amongst other things.
You can ignore this file if you prefer, but in my opinion it's better than SP1.